requirements.txt
langchain==0.3.13
langchain-openai==0.2
chromadb==0.5.13
azure-ai-documentintelligence==1.0.0
azure-core==1.32.0
python-dotenv==1.0.1
boto3==1.35.86
docx2txt==0.8
azure-cognitiveservices-vision-computervision==0.9.1
docling==2.14.0
pypdf==5.1.0
amazon-textract-caller==0.2.4
amazon-textract-textractor==1.8.5
azure-storage-blob==12.24.0
unstructured==0.15
llama-parse==0.5.18
openai==1.58.1
pdf2image==1.17.0
pinecone[grpc]==5.4.2
langchain-chroma==0.1.4
rank-bm25==0.2.2
langchain-community==0.3.13
pytest==8.3.2
mysqlclient==2.1.1
langchain-experimental==0.3.4
psycopg2-binary==2.9.10
pyodbc==5.2.0
sqlalchemy==2.0.36
beautifulsoup4
mammoth
streamlit
typing_extensions
nltk
langgraph
spacy